Finish swipes for user
even if the swipes do not finish exactly on a tile.  The OrphanSwipeCatcher only covered have the screen so swipes that did not end either on a tile or on the top half of the screen were not completed unless the user moved mouse back to last tile and clicked again.